+++ typescript/sdk/src/token/Token.test.ts
@@ -191,6 +191,7 @@ const STANDARD_TO_TOKEN: Record<TokenStandard, TokenArgs | null> = {
     symbol: 'SOL',
     name: 'SOL',
   },
+  [TokenStandard.SealevelHypCrossCollateral]: null,
 
   // Cosmos
   [TokenStandard.CosmosIcs20]: null,